Remove the concept of a GAE service endpoint
We don't need to support the mix of GAE and GKE any more so we can get
rid of the GaeService bits and unify everything under one constant
service. This also allows us to reduce the number of services down to
four (FE, BE, PUBAPI, console) which is nice.

For now we still keep around BackendRequestComponent and
FrontendRequestComponent -- they're only used by the test server, and we
should probably eventually find another way to serve those in local
testing.